PERF: nested dict DataFrame construction
